refactor(ch10): split a2a/ into a2a/client/ ahead of #787
Moves spec.py, tools.py, utils.py, constants.py into a2a/client/
(internal relative imports between them are unaffected -- they moved
together). a2a/__init__.py re-exports A2AAgentSpec/UseA2AAgentTool
from .client unchanged, so external imports of
llm_agents_from_scratch.a2a (including examples/ch10.ipynb) are
untouched; only internal absolute imports of the old submodule paths
(agent/builder.py, agent/llm_agent.py, and mock patch targets in
tests) needed updating.

Prep for #787 (LLMAgentA2AExecutor + card-building function), which
will land under a2a/server/ -- the serving side is a genuinely
different concern (implementing the SDK's AgentExecutor interface,
per-request event-queue publishing) from the client-dispatch code
here, not a guess at structure for code that doesn't exist yet.

tests/a2a/ mirrors the new src/ layout, consistent with
tests/tools/mcp/ mirroring src/.../tools/mcp/ elsewhere in the repo.
