+++ src/llm_agents_from_scratch/a2a/__init__.py
@@ -1,7 +1,6 @@
 """A2A module."""
 
-from .spec import A2AAgentSpec
-from .tools import UseA2AAgentTool
+from .client import A2AAgentSpec, UseA2AAgentTool
 
 __all__ = [
     "A2AAgentSpec",
