+++ src/llm_agents_from_scratch/a2a/client/spec.py
